requirements.txt
arrow==1.2.3
attrs==25.4.0
cachetools==5.2.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==2.0.12
cryptography==46.0.5
DateTime==4.7
dotenv==0.9.9
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.6.0
googleapis-common-protos==1.56.4
gspread==6.2.1
httplib2==0.22.0
icalendar==6.1.3
icalevents==0.2.1
idna==3.7
markdown-it-py==4.0.0
mdurl==0.1.2
oauthlib==3.2.2
protobuf==4.25.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==3.0
Pygments==2.19.2
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
requests==2.32.4
requests-oauthlib==1.3.1
rich==14.3.3
rsa==4.9
setuptools==82.0.0
six==1.16.0
TatSu==5.17.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.6.3
zope.interface==5.5.0
